Move dataSetDebugInformation to the separate file
